assignment_one/agent/requirements.txt
livekit-agents==1.2.6
livekit-plugins-openai==1.2.6
livekit-plugins-silero==1.2.6
python-dotenv==1.0.1
